Rename (_gen.go suffix) and split existing files for database role
